refactor(ICache): move ICacheMSHR into separate file
