+++ src/main/scala/xiangshan/frontend/icache/ICacheMissUnit.scala
@@ -25,93 +25,6 @@ import org.chipsalliance.cde.config.Parameters
 import utility._
 import xiangshan._
 
-class ICacheMSHRIO(edge: TLEdgeOut)(implicit p: Parameters) extends ICacheBundle {
-  val fencei:    Bool                       = Input(Bool())
-  val flush:     Bool                       = Input(Bool())
-  val invalid:   Bool                       = Input(Bool())
-  val req:       DecoupledIO[ICacheMissReq] = Flipped(DecoupledIO(new ICacheMissReq))
-  val acquire:   DecoupledIO[MSHRAcquire]   = DecoupledIO(new MSHRAcquire(edge))
-  val lookUps:   Vec[LookUpMSHR]            = Flipped(Vec(2, new LookUpMSHR))
-  val resp:      Valid[MSHRResp]            = ValidIO(new MSHRResp)
-  val victimWay: UInt                       = Input(UInt(wayBits.W))
-}
-
-class ICacheMSHR(edge: TLEdgeOut, isFetch: Boolean, ID: Int)(implicit p: Parameters) extends ICacheModule {
-  val io: ICacheMSHRIO = IO(new ICacheMSHRIO(edge))
-
-  private val valid = RegInit(Bool(), false.B)
-  // this MSHR doesn't respond to fetch and sram
-  private val flush  = RegInit(Bool(), false.B)
-  private val fencei = RegInit(Bool(), false.B)
-  // this MSHR has been issued
-  private val issue = RegInit(Bool(), false.B)
-
-  private val blkPaddr = RegInit(UInt((PAddrBits - blockOffBits).W), 0.U)
-  private val vSetIdx  = RegInit(UInt(idxBits.W), 0.U)
-  private val way      = RegInit(UInt(wayBits.W), 0.U)
-
-  // look up and return result at the same cycle
-  private val hits = io.lookUps.map { lookup =>
-    valid && !fencei && !flush && (lookup.info.bits.vSetIdx === vSetIdx) &&
-    (lookup.info.bits.blkPaddr === blkPaddr)
-  }
-  // Decoupling valid and bits
-  (0 until 2).foreach(i => io.lookUps(i).hit := hits(i))
-
-  // disable wake up when hit MSHR (fencei is low)
-  // when(hit) {
-  //   flush := false.B
-  // }
-
-  // invalid when the req hasn't been issued
-  when(io.fencei || io.flush) {
-    fencei := true.B
-    flush  := true.B
-    when(!issue) {
-      valid := false.B
-    }
-  }
-
-  // receive request and register
-  io.req.ready := !valid && !io.flush && !io.fencei
-  when(io.req.fire) {
-    valid    := true.B
-    flush    := false.B
-    issue    := false.B
-    fencei   := false.B
-    blkPaddr := io.req.bits.blkPaddr
-    vSetIdx  := io.req.bits.vSetIdx
-  }
-
-  // send request to L2
-  io.acquire.valid := valid && !issue && !io.flush && !io.fencei
-  private val getBlock = edge.Get(
-    fromSource = ID.U,
-    toAddress = Cat(blkPaddr, 0.U(blockOffBits.W)),
-    lgSize = log2Up(cacheParams.blockBytes).U
-  )._2
-  io.acquire.bits.acquire := getBlock
-  io.acquire.bits.acquire.user.lift(ReqSourceKey).foreach(_ := MemReqSource.CPUInst.id.U)
-  io.acquire.bits.vSetIdx := vSetIdx
-
-  // get victim way when acquire fire
-  when(io.acquire.fire) {
-    issue := true.B
-    way   := io.victimWay
-  }
-
-  // invalid request when grant finish
-  when(io.invalid) {
-    valid := false.B
-  }
-
-  // offer the information other than data for write sram and response fetch
-  io.resp.valid         := valid && (!flush && !fencei)
-  io.resp.bits.blkPaddr := blkPaddr
-  io.resp.bits.vSetIdx  := vSetIdx
-  io.resp.bits.way      := way
-}
-
 class ICacheMissUnitIO(edge: TLEdgeOut)(implicit p: Parameters) extends ICacheBundle {
   // difftest
   val hartId: Bool = Input(Bool())
